Split autostart implementation in two different files
