+++ CPicker/CPicker.vcxproj
@@ -1,14 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-
   <PropertyGroup Label="Globals">
     <ProjectGuid>{AB4D3015-2AD4-4152-BDD2-FC1343B22B6C}</ProjectGuid>
     <RootNamespace>CPicker</RootNamespace>
     <ConfigurationType>DynamicLibrary</ConfigurationType>
   </PropertyGroup>
-
   <Import Project="..\common.props" />
-
   <ItemDefinitionGroup Label="Globals">
     <ClCompile>
       <PreprocessorDefinitions>_CPICKER_DLL;%(PreprocessorDefinitions)</PreprocessorDefinitions>
@@ -18,7 +15,6 @@
       <AdditionalManifestDependencies>"type='win32' name='Microsoft.Windows.Common-Controls' version='6.0.0.0' processorArchitecture='*' publicKeyToken='6595b64144ccf1df' language='*'"</AdditionalManifestDependencies>
     </Link>
   </ItemDefinitionGroup>
-
   <ItemGroup>
     <ClCompile Include="alphaslidercontext.cpp" />
     <ClCompile Include="ccolourpicker.cpp" />
