+++ TranslucentTB/TranslucentTB.vcxproj
@@ -1,23 +1,18 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-
   <PropertyGroup Label="Globals">
     <ProjectGuid>{59F844AA-8D3C-431C-B8CC-57682915F551}</ProjectGuid>
     <RootNamespace>TranslucentTB</RootNamespace>
     <ConfigurationType>Application</ConfigurationType>
   </PropertyGroup>
-
   <Import Project="..\common.props" />
-
   <ItemDefinitionGroup Label="Globals">
     <Link>
       <AdditionalDependencies>advapi32.lib;comctl32.lib;dwmapi.lib;ole32.lib;pathcch.lib;runtimeobject.lib;shcore.lib;shell32.lib;user32.lib</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
-
   <ItemGroup>
     <ClCompile Include="appvisibilitysink.cpp" />
-    <ClCompile Include="autostart.cpp" />
     <ClCompile Include="blacklist.cpp" />
     <ClCompile Include="config.cpp" />
     <ClCompile Include="eventhook.cpp" />
@@ -57,6 +52,18 @@
     <ClInclude Include="window.hpp" />
     <ClInclude Include="windowclass.hpp" />
   </ItemGroup>
+  <Choose>
+    <When Condition="'$(Configuration)'=='Store'">
+      <ItemGroup>
+        <ClCompile Include="autostart_store.cpp" />
+      </ItemGroup>
+    </When>
+    <Otherwise>
+      <ItemGroup>
+        <ClCompile Include="autostart_desktop.cpp" />
+      </ItemGroup>
+    </Otherwise>
+  </Choose>
   <ItemGroup Condition="'$(Configuration)'=='Store'">
     <ClCompile Include="uwp.cpp" />
     <ClInclude Include="uwp.hpp" />
